requirements.txt
google-cloud-language==2.9.0
numpy==1.24.2
pandas==1.5.3
six==1.16.0
vaderSentiment==3.3.2
